+++ android/src/main/java/com/cloudwebrtc/webrtc/FlutterWebRTCPlugin.java
@@ -7,6 +7,7 @@ import android.os.Bundle;
 import android.util.Log;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.lifecycle.DefaultLifecycleObserver;
 import androidx.lifecycle.Lifecycle;
 import androidx.lifecycle.LifecycleOwner;
@@ -19,6 +20,7 @@ import com.cloudwebrtc.webrtc.utils.ConstraintsMap;
 import org.webrtc.ExternalAudioProcessingFactory;
 import org.webrtc.MediaStreamTrack;
 import org.webrtc.PeerConnectionFactory;
+import org.webrtc.audio.JavaAudioDeviceModule;
 
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.embedding.engine.plugins.activity.ActivityAware;
@@ -72,6 +74,11 @@ public class FlutterWebRTCPlugin implements FlutterPlugin, ActivityAware, EventC
 
     public PeerConnectionFactory getPeerConnectionFactory() {
         return methodCallHandler.getPeerConnectionFactory();
+    }
+
+    @Nullable
+    public JavaAudioDeviceModule getAudioDeviceModule() {
+        return methodCallHandler == null ? null : methodCallHandler.getAudioDeviceModule();
     }
 
     @Override